Version 0.12 - Incomplete
Worked on a lot of different things in this update. Some including:
Improved error checking,
Removed switch case statements,
Made it simpler to use single line statements,
Added a token called BOOL_REVERSE.

This commit is incomplete, so do not use this commit if you want a
stable version of the project.