Version 0.12g - Rewriting the syntax checker
The lexer has been fully rewritten and is now complete*. The next stage is to rewrite the syntax checker to include something similar to a syntax tree.

For example, if the starting token of a line is "FUNCTION_CALL", then the next token must either be "IDENTIFIER" or "LINEFEED". If it is not correct, an error is produced.

* The lexer may be updated in the future.